+++ src/services/CacheRequestService.php
@@ -22,7 +22,6 @@ use putyourlightson\blitz\helpers\SiteUriHelper;
 use putyourlightson\blitz\models\SettingsModel;
 use putyourlightson\blitz\models\SiteUriModel;
 use putyourlightson\blitz\records\IncludeRecord;
-use putyourlightson\blitz\variables\BlitzVariable;
 use yii\web\Response;
 
 /**
@@ -52,7 +51,22 @@ class CacheRequestService extends Component
     /**
      * @const string
      */
+    public const CACHED_INCLUDE_ACTION = 'blitz/include/cached';
+
+    /**
+     * @const string
+     */
+    public const DYNAMIC_INCLUDE_ACTION = 'blitz/include/dynamic';
+
+    /**
+     * @const string
+     */
     public const CACHED_INCLUDE_PATH = '_includes';
+
+    /**
+     * @const string
+     */
+    public const DYNAMIC_INCLUDE_PATH = '_dynamic';
 
     /**
      * @var bool|null
@@ -146,6 +160,10 @@ class CacheRequestService extends Component
             return true;
         }
 
+        if ($this->getIsDynamicInclude()) {
+            return false;
+        }
+
         return $response->format == Response::FORMAT_HTML
             || $response->format == 'template'
             || Blitz::$plugin->settings->cacheNonHtmlResponses;
@@ -168,6 +186,10 @@ class CacheRequestService extends Component
 
         if ($this->getIsCachedInclude($uri)) {
             return true;
+        }
+
+        if ($this->getIsDynamicInclude($uri)) {
+            return false;
         }
 
         $url = $siteUri->getUrl();
@@ -238,7 +260,31 @@ class CacheRequestService extends Component
         if (Craft::$app->getRequest()->getIsActionRequest()) {
             $action = implode('/', Craft::$app->getRequest()->getActionSegments());
 
-            return $action == BlitzVariable::CACHED_INCLUDE_ACTION;
+            return $action == self::CACHED_INCLUDE_ACTION;
+        }
+
+        return false;
+    }
+
+    /**
+     * Returns whether this is a dynamic include.
+     * Doesn’t memoize the result, which would disrupt the local cache generator.
+     *
+     * @since 4.6.0
+     */
+    public function getIsDynamicInclude(string $uri = null): bool
+    {
+        // Includes based on the URI takes preference
+        if ($uri !== null) {
+            $uri = trim($uri, '/');
+
+            return str_starts_with($uri, self::DYNAMIC_INCLUDE_PATH);
+        }
+
+        if (Craft::$app->getRequest()->getIsActionRequest()) {
+            $action = implode('/', Craft::$app->getRequest()->getActionSegments());
+
+            return $action == self::DYNAMIC_INCLUDE_ACTION;
         }
 
         return false;